+++ resources/lib/model/fetcher.py
@@ -5,12 +5,16 @@ class Fetcher:
     def __init__(self, logger):
         self.logger = logger
 
-    def get_HTML(self, url, data=None):
+    def get(self, url):
         self.logger('GET {}'.format(url))
 
         r = requests.get(url)
         if r.ok:
-            return r.text
+            return r
         else:
             self.logger('%s\n%s\n%s'.format(r.status_code, r.headers, r.text))
             raise Exception('Failed to GET {}: {}'.format(url, r.status_code))
+
+
+    def get_HTML(self, url):
+        return self.get(url).text