+++ resources/lib/model/talk_scraper.py
@@ -3,43 +3,36 @@ import json
 import requests
 
 
-def get(html, logger, video_quality='180kbps'):
+def get_talk(html, logger):
     '''Extract talk details from talk html
-       @param video_quality string in form '\\d+kbps' that should match one of the provided TED bitrates.
     '''
 
     init_scripts = html5lib.parse(html, namespaceHTMLElements=False).findall(".//script[@data-spec='q']")
     init_scripts = [script.text for script in init_scripts if '"talkPage.init"' in script.text]
-    if init_scripts:
-        # let's do this some other way
-        init_scripts_start_pos = str(init_scripts).find('{')
-        init_scripts = str(init_scripts)[init_scripts_start_pos:]
-        init_scripts = init_scripts + '}'
 
-        # let's remove some json breaking invalid characters
-        init_scripts = init_scripts.replace('\\n\\t', '')
-        init_scripts = init_scripts.replace("\\n})']", '')
-        init_scripts = init_scripts.replace('\\\\"', '')
-        init_scripts = init_scripts.replace('\\', '')
-        init_scripts = init_scripts[:-4]
+    if not init_scripts:
+        raise Exception('Could not parse HTML.')
 
-        init_json = json.loads(str(init_scripts), strict=False)
-        talk_json = init_json['__INITIAL_DATA__']['talks'][0]
-        
-        title = talk_json['player_talks'][0]['title']
-        speaker = talk_json['player_talks'][0]['speaker']
-        plot = talk_json['description']
+    # let's do this some other way
+    init_scripts_start_pos = str(init_scripts).find('{')
+    init_scripts = str(init_scripts)[init_scripts_start_pos:]
+    init_scripts = init_scripts + '}'
 
-        url = talk_json['downloads']['nativeDownloads']['medium']
-        url = url[0:url.find('?')]
+    # let's remove some json breaking invalid characters
+    init_scripts = init_scripts.replace('\\n\\t', '')
+    init_scripts = init_scripts.replace("\\n})']", '')
+    init_scripts = init_scripts.replace('\\\\"', '')
+    init_scripts = init_scripts.replace('\\', '')
+    init_scripts = init_scripts[:-4]
 
-        # Attempt to switch to user-specified bitrate.
-        url_custom = url.replace('.mp4', '-%sk.mp4' % (video_quality.split('k')[0]))
-        if requests.head(url_custom).ok:
-            url = url_custom
-        else:
-            logger("Could not create URL for bitrate '%s' from '%s'." % (video_quality, url))
+    init_json = json.loads(str(init_scripts), strict=False)
+    talk_json = init_json['__INITIAL_DATA__']['talks'][0]
+    player_talks_json = talk_json['player_talks'][0]
+    
+    title = player_talks_json['title']
+    speaker = player_talks_json['speaker']
+    plot = talk_json['description']
 
-        return url, title, speaker, plot, talk_json
-    else:
-        raise Exception('Could not parse HTML:\n%s' % (html))
+    m3u8_url = player_talks_json['resources']['hls']['stream']
+
+    return m3u8_url, title, speaker, plot, talk_json