+++ resources/lib/settings.py
@@ -12,23 +12,17 @@ __current_search__ = 'current_search'
 __current_search_results__ = 'current_search_results'
 
 profile_path = '~/.xbmc/userdata/addon_data/plugin.video.ted.talks'
-download_mode = True
-download_path = '/tmp/'
-video_quality = 3
 enable_subtitles = True
 xbmc_language = 'English'
 subtitle_language = 'en'
 
 def init():
-    import xbmc, xbmcaddon
+    import xbmc, xbmcvfs, xbmcaddon
     addon = xbmcaddon.Addon(id=__plugin_id__)
-    global profile_path, download_mode, download_path, video_quality, enable_subtitles, xbmc_language, subtitle_language
-    profile_path = xbmc.translatePath(addon.getAddonInfo('profile')).decode("utf-8")
+    global profile_path, enable_subtitles, xbmc_language, subtitle_language
+    profile_path = xbmcvfs.translatePath(addon.getAddonInfo('profile'))
     if not os.path.exists(profile_path):
         os.makedirs(profile_path)
-    download_mode = addon.getSetting('downloadMode')
-    download_path = addon.getSetting('downloadPath')
-    video_quality = addon.getSetting('video_quality')
     enable_subtitles = addon.getSetting('enable_subtitles')
     xbmc_language = xbmc.getLanguage()
     subtitle_language = addon.getSetting('subtitle_language')